Add files via upload
The Hostel Management System is a comprehensive web application developed using Django, a Python-based framework. The front-end of the system utilizes HTML5, CSS3, and JavaScript to create an intuitive and visually appealing user interface. Django is employed for the back-end, providing a robust and scalable foundation for handling various functionalities. The SQLite-3 database is used to efficiently store and manage data.

The Hostel Management System aims to streamline and automate the management processes of a hostel. It offers a range of features to effectively handle staff members, residents, room allocation, food routines, and leave and outing permissions, providing a centralized platform for efficient administration and enhanced resident experiences.